took liquid material class out of the material file
